+++ .idea/sqldialects.xml
@@ -3,7 +3,7 @@
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/src/main/java/profchoper/_misc/DBTest.java" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/src/main/java/profchoper/course/CourseRepository.java" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/src/main/java/profchoper/student/StudentRepository.java" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/src/main/java/profchoper/student/StudentService.java" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/src/main/resources/data.sql" dialect="PostgreSQL" />
     <file url="file://$PROJECT_DIR$/src/main/resources/schema.sql" dialect="PostgreSQL" />
   </component>
