+++ src/main/java/profchoper/course/Course.java
@@ -2,13 +2,29 @@ package profchoper.course;
 
 
 public class Course {
-    private final String id;
-    private final String name;
-    private final String alias;
+    private String id;
+    private String name;
+    private String alias;
+
+    public Course() {
+        // empty constructor
+    }
 
     public Course(String id, String name, String alias) {
         this.id = id;
         this.name = name;
+        this.alias = alias;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setAlias(String alias) {
         this.alias = alias;
     }
