+++ src/main/resources/templates/index.html
@@ -31,7 +31,7 @@
 
 
     <!--/*@thymesVar id="courses" type="java.util.List"*/-->
-<!--    <div th:if="not ${courses.isEmpty()}">
+    <div th:if="not ${courses.isEmpty()}">
         <h1>Courses</h1>
         <table class="table table-striped">
             <tr>
@@ -40,17 +40,17 @@
                 <th>Alias</th>
             </tr>
 
-            &lt;!&ndash;/*@thymesVar id="course" type="profchoper.course.Course"*/&ndash;&gt;
+            <!--/*@thymesVar id="course" type="profchoper.course.Course"*/-->
             <tr th:each="course : ${courses}">
                 <td th:text="${course.getId()}"></td>
                 <td th:text="${course.getName()}"></td>
                 <td th:text="${course.getAlias()}"></td>
             </tr>
         </table>
-    </div>-->
+    </div>
 
     <!--/*@thymesVar id="students" type="java.util.List"*/-->
-    <div th:if="not ${students.isEmpty()}">
+<!--    <div th:if="not ${students.isEmpty()}">
         <h1>Students</h1>
         <table class="table table-striped">
             <tr>
@@ -59,7 +59,7 @@
                 <th>Email</th>
             </tr>
 
-            <!--/*@thymesVar id="student" type="profchoper.student.Student"*/-->
+            &lt;!&ndash;/*@thymesVar id="student" type="profchoper.student.Student"*/&ndash;&gt;
             <tr th:each="student : ${students}">
                 <td th:text="${student.getId()}"></td>
                 <td th:text="${student.getName()}"></td>
@@ -67,7 +67,7 @@
             </tr>
 
         </table>
-    </div>
+    </div>-->
 </div>
 
 </body>
